refactor: moves the utils
